move Unity configuration code to UnityConfig class
