+++ Bonobo.Git.Server/Global.asax.cs
@@ -2,20 +2,12 @@
 using Bonobo.Git.Server.Attributes;
 using Bonobo.Git.Server.Configuration;
 using Bonobo.Git.Server.Controllers;
-using Bonobo.Git.Server.Data;
 using Bonobo.Git.Server.Data.Update;
-using Bonobo.Git.Server.Git;
-using Bonobo.Git.Server.Git.GitService;
-using Bonobo.Git.Server.Git.GitService.ReceivePackHook;
-using Bonobo.Git.Server.Git.GitService.ReceivePackHook.Durability;
-using Bonobo.Git.Server.Git.GitService.ReceivePackHook.Hooks;
-using Bonobo.Git.Server.Security;
 using Serilog;
 using System;
 using System.Configuration;
 using System.Globalization;
 using System.IO;
-using System.Linq;
 using System.Runtime.Caching;
 using System.Security.Claims;
 using System.Threading;
@@ -26,10 +18,6 @@ using System.Web.Hosting;
 using System.Web.Mvc;
 using System.Web.Optimization;
 using System.Web.Routing;
-using Unity;
-using Unity.AspNet.Mvc;
-using Unity.Injection;
-using Unity.Resolution;
 
 namespace Bonobo.Git.Server
 {
@@ -79,9 +67,10 @@ namespace Bonobo.Git.Server
             AreaRegistration.RegisterAllAreas();
             BundleConfig.RegisterBundles(BundleTable.Bundles);
             RouteConfig.RegisterRoutes(RouteTable.Routes);
+            UnityConfig.RegisterFilters();
+
             UserConfiguration.Initialize();
-            RegisterDependencyResolver();
-            GlobalFilters.Filters.Add((AllViewsFilter)DependencyResolver.Current.GetService<AllViewsFilter>());
+            GlobalFilters.Filters.Add(DependencyResolver.Current.GetService<AllViewsFilter>());
 
             var connectionstring = WebConfigurationManager.ConnectionStrings["BonoboGitServerContext"];
             if (connectionstring.ProviderName.ToLowerInvariant() == "system.data.sqlite")
@@ -123,132 +112,6 @@ namespace Bonobo.Git.Server
                 logDirectory = @"~\App_Data\Logs";
             }
             return Path.Combine(HostingEnvironment.MapPath(logDirectory), "log-{Date}.txt");
-        }
-
-        private static void RegisterDependencyResolver()
-        {
-            var container = new UnityContainer();
-
-            /* 
-                The UnityDecoratorContainerExtension breaks resolving named type registrations, like:
-
-                container.RegisterType<IMembershipService, ADMembershipService>("ActiveDirectory");
-                container.RegisterType<IMembershipService, EFMembershipService>("Internal");
-                IMembershipService membershipService = container.Resolve<IMembershipService>(AuthenticationSettings.MembershipService);
-
-                Until this issue is resolved, the following two switch hacks will have to do
-            */
-
-            switch (AuthenticationSettings.MembershipService.ToLowerInvariant())
-            {
-                case "activedirectory":
-                    container.RegisterType<IMembershipService, ADMembershipService>();
-                    container.RegisterType<IRoleProvider, ADRoleProvider>();
-                    container.RegisterType<ITeamRepository, ADTeamRepository>();
-                    container.RegisterType<IRepositoryRepository, ADRepositoryRepository>();
-                    container.RegisterType<IRepositoryPermissionService, RepositoryPermissionService>();
-                    break;
-                case "internal":
-                    container.RegisterType<IMembershipService, EFMembershipService>();
-                    container.RegisterType<IRoleProvider, EFRoleProvider>();
-                    container.RegisterType<ITeamRepository, EFTeamRepository>();
-                    container.RegisterType<IRepositoryRepository, EFRepositoryRepository>();
-                    container.RegisterType<IRepositoryPermissionService, RepositoryPermissionService>();
-                    break;
-                default:
-                    throw new ArgumentException("Missing declaration in web.config", "MembershipService");
-            }
-
-            switch (AuthenticationSettings.AuthenticationProvider.ToLowerInvariant())
-            {
-                case "windows":
-                    container.RegisterType<IAuthenticationProvider, WindowsAuthenticationProvider>();
-                    break;
-                case "cookies":
-                    container.RegisterType<IAuthenticationProvider, CookieAuthenticationProvider>();
-                    break;
-                case "federation":
-                    container.RegisterType<IAuthenticationProvider, FederationAuthenticationProvider>();
-                    break;
-                default:
-                    throw new ArgumentException("Missing declaration in web.config", "AuthenticationProvider");
-            }
-
-            container.RegisterFactory<IGitRepositoryLocator>((ctr, type, name) => new ConfigurationBasedRepositoryLocator(UserConfiguration.Current.Repositories));
-
-            container.RegisterInstance(
-                new GitServiceExecutorParams()
-                {
-                    GitPath = GetRootPath(ConfigurationManager.AppSettings["GitPath"]),
-                    GitHomePath = GetRootPath(ConfigurationManager.AppSettings["GitHomePath"]),
-                    RepositoriesDirPath = UserConfiguration.Current.Repositories,
-                });
-
-            container.RegisterType<IDatabaseResetManager, DatabaseResetManager>();
-
-            if (AppSettings.IsPushAuditEnabled)
-            {
-                EnablePushAuditAnalysis(container);
-            }
-
-            container.RegisterType<IGitService, GitServiceExecutor>();
-
-            DependencyResolver.SetResolver(new UnityDependencyResolver(container));
-
-            var oldProvider = FilterProviders.Providers.Single(f => f is FilterAttributeFilterProvider);
-            FilterProviders.Providers.Remove(oldProvider);
-
-            var provider = new UnityFilterAttributeFilterProvider(container);
-            FilterProviders.Providers.Add(provider);
-        }
-
-        private static void EnablePushAuditAnalysis(IUnityContainer container)
-        {
-            var isReceivePackRecoveryProcessEnabled = !string.IsNullOrEmpty(ConfigurationManager.AppSettings["RecoveryDataPath"]);
-
-            if (isReceivePackRecoveryProcessEnabled)
-            {
-                // git service execution durability registrations to enable receive-pack hook execution after failures
-                container.RegisterType<IGitService, DurableGitServiceResult>();
-                container.RegisterType<IHookReceivePack, ReceivePackRecovery>();
-                container.RegisterType<IRecoveryFilePathBuilder, AutoCreateMissingRecoveryDirectories>();
-                container.RegisterType<IRecoveryFilePathBuilder, OneFolderRecoveryFilePathBuilder>();
-                container.RegisterInstance(new NamedArguments.FailedPackWaitTimeBeforeExecution(TimeSpan.FromSeconds(5 * 60)));
-
-                container.RegisterInstance(new NamedArguments.ReceivePackRecoveryDirectory(
-                    Path.IsPathRooted(ConfigurationManager.AppSettings["RecoveryDataPath"]) ?
-                        ConfigurationManager.AppSettings["RecoveryDataPath"] :
-                        HttpContext.Current.Server.MapPath(ConfigurationManager.AppSettings["RecoveryDataPath"])));
-            }
-
-            // base git service executor
-            container.RegisterType<IGitService, ReceivePackParser>();
-            container.RegisterType<GitServiceResultParser, GitServiceResultParser>();
-
-            // receive pack hooks
-            container.RegisterType<IHookReceivePack, AuditPusherToGitNotes>();
-            container.RegisterType<IHookReceivePack, NullReceivePackHook>();
-
-            // run receive-pack recovery if possible
-            if (isReceivePackRecoveryProcessEnabled)
-            {
-                var recoveryProcess = container.Resolve<ReceivePackRecovery>(
-                    new ParameterOverride(
-                        "failedPackWaitTimeBeforeExecution",
-                        new NamedArguments.FailedPackWaitTimeBeforeExecution(TimeSpan.FromSeconds(0)))); // on start up set time to wait = 0 so that recovery for all waiting packs is attempted
-
-                try
-                {
-                    recoveryProcess.RecoverAll();
-                }
-                catch
-                {
-                    // don't let a failed recovery attempt stop start-up process
-                }
-                finally
-                {
-                }
-            }
         }
 
 
@@ -293,13 +156,5 @@ namespace Bonobo.Git.Server
                 errorController.Execute(new RequestContext(new HttpContextWrapper(Context), routeData));
             }
         }
-
-        private static string GetRootPath(string path)
-        {
-            return Path.IsPathRooted(path) ?
-                path :
-                HttpContext.Current.Server.MapPath(path);
-        }
-
     }
 }
